fix: ensure stable imports
